+++ NEWS.md
@@ -1,5 +1,7 @@
 # provParseR 0.1.2
 
+* Added functions: get.urls, get.error.nodes, and get.variable.named
+* Added only.files parameter to get.input.files.
 * Fixed bug in the parsing of sourced scripts
 * Changed the getter functions so that when they return an empty data frame, it still has column names